Adding server frame validation
As per specification, frames coming from the server are now validated
as well.

Unfortunately, there is one validation for the Message frame that is
conditional and based on state (a previous message sent to the server).

I designed this framework to serve as a foundation for Stomp messages
in any Websocket communication, so storing state is a no go. That logic
will need to go one level higher and should be handled by whoever is
using this framework.