Refactoring the interface for "bodyless" frames  to be simpler
Most of the frames cannot have a body as per specification. So long
the interface of the Frames allowed devs to pass any arbitrary value for
the body, and there was a runtime validation to decide whether it's allowed
or not.

Now the constructor of these Frames don't expect a body, hence there's no way to
provide it.